feat: 7670 - adding a single test file as CreateCohortDistricutionDataTests
