+++ tests/UnitTests/CohortDistributionTests/CohortDistributionDataTests/CreateCohortDistributionDataTests.cs
@@ -11,7 +11,7 @@ using System.Collections.Generic;
 using System.Linq;
 
 [TestClass]
-public class AddCohortDistributionTests
+public class CreateCohortDistributionDataTests
 {
     private readonly Mock<IDbConnection> _mockDBConnection = new();
     private readonly Mock<IDbCommand> _commandMock = new();
@@ -21,9 +21,11 @@ public class AddCohortDistributionTests
     private readonly Mock<IDbTransaction> _mockTransaction = new();
     private readonly Mock<IDataServiceClient<CohortDistribution>> _cohortDistributionMock = new();
     private readonly Mock<IDataParameterCollection> _parameterCollectionMock = new();
+     protected CreateCohortDistributionData _createCohortDistributionDataService;
 
 
-    public AddCohortDistributionTests()
+
+    public CreateCohortDistributionDataTests()
     {
         _mockDBConnection.Setup(x => x.ConnectionString).Returns("someFakeConnectionString");
         _mockDBConnection.Setup(x => x.BeginTransaction()).Returns(_mockTransaction.Object);
@@ -31,6 +33,13 @@ public class AddCohortDistributionTests
         _commandMock.Setup(m => m.CreateParameter()).Returns(new Mock<IDbDataParameter>().Object);
         _commandMock.Setup(m => m.Parameters).Returns(new Mock<IDataParameterCollection>().Object);
         _commandMock.Setup(m => m.Parameters).Returns(_parameterCollectionMock.Object);
+
+                _mockDBConnection.Setup(m => m.CreateCommand()).Returns(_commandMock.Object);
+        _commandMock.Setup(m => m.CreateParameter()).Returns(new Mock<IDbDataParameter>().Object);
+        _commandMock.Setup(m => m.Parameters).Returns(new Mock<IDataParameterCollection>().Object);
+        _commandMock.Setup(m => m.ExecuteReader()).Returns(_mockDataReader.Object);
+
+
 
     }
 
@@ -395,6 +404,30 @@ public class AddCohortDistributionTests
         Assert.IsNotNull(result, "Result should not be null");
         Assert.IsTrue(string.IsNullOrEmpty(result.RequestId), "RequestId should be empty");
         Assert.IsTrue(string.IsNullOrEmpty(result.StatusCode), "StatusCode should be empty");
+    }
+
+    [TestMethod]
+    public void UpdateCohortDistributionParticipantAsInactive_Success()
+    {
+        // Arrange
+        var updateCohortDistributionData = new CreateCohortDistributionData(
+            _mockDBConnection.Object,
+            _loggerMock.Object,
+            _cohortDistributionMock.Object
+        );
+
+        string NHSID = "123456";
+
+        _commandMock.Setup(m => m.ExecuteNonQuery()).Returns(1);
+        _commandMock.Setup(m => m.CommandText).Returns("UPDATE Cohort SET Status='Inactive' WHERE NHSID=@NHSID");
+        _commandMock.Setup(m => m.Parameters).Returns(new Mock<IDataParameterCollection>().Object);
+
+        // Act
+        var result = updateCohortDistributionData.UpdateCohortParticipantAsInactive(NHSID);
+
+        // Assert
+        Assert.IsTrue(result);
+        _commandMock.Verify(m => m.ExecuteNonQuery(), Times.Once);
     }
 
 }
